Separated some concerns so that there are specific classes for specific actions - this is reflected in the tests
